pemisahan package gui dan logic
biar enak
