+++ ppdbSma/nbproject/private/private.xml
@@ -2,13 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/workspaces/public-github/Tugas-Linked-List-TI-2B/ppdbSma/src/ppdbSma/MainLogic.java</file>
-            <file>file:/C:/workspaces/public-github/Tugas-Linked-List-TI-2B/ppdbSma/src/ppdbSma/ProsesSeleksi.java</file>
-            <file>file:/C:/workspaces/public-github/Tugas-Linked-List-TI-2B/ppdbSma/src/ppdbSma/FormSiswaPilih.java</file>
-            <file>file:/C:/workspaces/public-github/Tugas-Linked-List-TI-2B/ppdbSma/src/ppdbSma/FrameLogin.java</file>
-            <file>file:/C:/workspaces/public-github/Tugas-Linked-List-TI-2B/ppdbSma/src/ppdbSma/ModelTabel.java</file>
-            <file>file:/C:/workspaces/public-github/Tugas-Linked-List-TI-2B/ppdbSma/src/ppdbSma/Randomizer.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
